Add live outputs from the loop tail to the loop params
Added a test that goes from failing to passing with this change. In this case
the `ans` variable is overwritten in the loop body but it is still used outside
so it needs to be given an input and output.

Note that for variables that are truly unbound at this point we may have to add
a placeholder variable to get them working.